fixed some issues and new feature for FindItem
there was some issues with the shiftlock stuff, for example. it was 

if (Gdip_ImageSearch((pBMScreen := Gdip_BitmapFromScreen(windowX+5 "|" windowY+windowHeight-54 "|50|50")), bitmaps["shiftlock"], , , , , , 2)) {
	send {%SC_LShift%}
}

but now i changed it to
if (Gdip_ImageSearch((pBMScreen := Gdip_BitmapFromScreen(windowX+5 "|" windowY+windowHeight-54 "|50|50")), bitmaps["shiftlock"], , , , , , 2)) {
	global localSC_LShift
	send {%localSC_LShift%}
	WasShiftlockEnabled:=True
}
(little hint of what i added)
i made it so it re-enables shiftlock if it was activated previously
if (WasShiftlockEnabled) {
	Send {%localSC_LShift%}
}
tell me if there was a better way of doing this. but there is also another fix with pausing, still there is one issue when unpausing, but fixed the issue where it doesn't display the keybind (same with the remote control logs) .