StickerStackVoucher implementation + Sticker Stack bug fix
StickerStackVoucher - This was a simple addition for the possibility of donating vouchers. "I just needed it"

Sticker Stack bug fix
the lines were off, this now is the correct branch. sorry if I have caused any difficulties.

inside natro_macro.ahk

line 11376 - increased the x-axis parameter to include the 4th row of stickers, why was it only configured to only do 3 ?

line 11377 & 11381- increased outer y-axis image search detection since 25 was not enough "cuts off half way from the text", increased by 20 pixels to get a bigger degree of searching

line 11402-11405 - added the voucher detection, plus made the variation from 15 to 25 since in roblox, the images have a grain, so this was messed up for the "grey" parts, so only white pixels are in the bitmap.

line 11436 - added a part where it's for the voucher, since voucher is not a "skin".

inside status.ahk
line 642 - added "StickerStackVoucher" to element 364 so the person can remotely disable it

inside stickerstack/bitmaps.ahl
line 14- 24
left in the old bitmaps for safe keeping if needed, and added my own bitmaps for "normal" and "voucher" since I just needed those 2 fixed to donate stickers I needed to donate.

line 25 - trailing newline

inside EnumInt.ahk
line 363 - not important but indentation was wack
line 364 - StickerStackVoucher element for Status.ahk

there are other changes like GUI implementation but that's another can of worms.