Add a new hive check system
SetHiveCameraDirection:
allows to search for the hive anchor for a certain amount of rotations.
Note: it is not added in the test pattern as it is just a copy of what's inside the reset function.
Also added nm_convert so that even if the hive check fails, pollen inside the player's bag will still be converted